% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_conclusion.R
\name{auto_conclusion}
\alias{auto_conclusion}
\title{Automatically generates a written conclusion for a regression model}
\usage{
auto_conclusion(model)
}
\arguments{
\item{model}{A fitted lm() object}
}
\value{
A character string summarizing model interpretation
}
\description{
Automatically generates a written conclusion for a regression model
}
