% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model.R
\name{check_model}
\alias{check_model}
\title{Check a regression model for issues}
\usage{
check_model(model)
}
\arguments{
\item{model}{An lm model object}
}
\value{
A list containing diagnostics and warnings
}
\description{
Runs diagnostics on an lm model including multicollinearity,
residual normality, heteroskedasticity, leverage, and influence.
}
\examples{
model <- lm(mpg ~ hp + wt, data = mtcars)
check_model(model)

}
