% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report_model}
\alias{report_model}
\title{Generate an HTML regression report}
\usage{
report_model(model, file = "model_report.html")
}
\arguments{
\item{model}{An lm model object.}

\item{file}{Output HTML file name.}
}
\value{
Path to created HTML file.
}
\description{
Generate an HTML regression report
}
